Navbar modification
The anchors elements were link to each section and the transition to one section to another was implemented with SmoothScroll js script got in github. The file smooth-scroll.js was modified in the line 311aprox. (var getHeaderHeight = function ( header ) {...) making it return always 46 which is the height of the navbar, if I wouldn't have done this change, the final position would have been wrong because the navbar block the top of the section